+++ src/children/index.js
@@ -1,6 +1,8 @@
+import { AssetLink, PartyLink, TransactionLink, Link } from './Link'
+
 export { default as Code } from './Code'
 export { default as Charge } from './Charge'
 export { default as Comment } from './Comment'
-export { default as Link } from './Link'
 export { default as Party } from './Party'
 export { default as Reference } from './Reference'
+export { AssetLink, PartyLink, TransactionLink, Link }