+++ src/index.js
@@ -4,7 +4,7 @@ import { books } from './exports/books'
 import { children } from './exports/children'
 import { corporateActions } from './exports/corporateActions'
 import { parties } from './exports/parties'
-import * as transactions from './exports/transactions'
+import { transactions } from './exports/transactions'
 
 import * as api from './exports/api'