+++ src/parties/Party/party.test.js
@@ -1,6 +1,6 @@
 import Party from './party.js'
 import { Address, Email } from '../Children'
-import { Comment, Link, Reference } from '../../children'
+import { Comment, PartyLink, Reference } from '../../children'
 
 describe('Party', () => {
   describe('serialization', () => {
@@ -69,21 +69,21 @@ describe('Party', () => {
     it('should set links correctly', () => {
       const party = new Party({})
       party.links = {
-        SINGLE1: [{ linkedId: 'ID-S1-1' }],
-        SINGLE2: [new Link({ linkedId: 'ID-S2-1' })],
+        SINGLE1: [{ linkedPartyId: 'ID-S1-1' }],
+        SINGLE2: [new PartyLink({ linkedPartyId: 'ID-S2-1' })],
         MULTI1: [
-          { linkedId: 'ID-M1-1' },
-          new Link({ linkedId: 'ID-M1-2' })
+          { linkedPartyId: 'ID-M1-1' },
+          new PartyLink({ linkedPartyId: 'ID-M1-2' })
         ]
       }
       expect(party.links.SINGLE1[0]).toBeDefined()
-      expect(party.links.SINGLE1[0].linkedId).toEqual('ID-S1-1')
+      expect(party.links.SINGLE1[0].linkedPartyId).toEqual('ID-S1-1')
       expect(party.links.SINGLE2[0]).toBeDefined()
-      expect(party.links.SINGLE2[0].linkedId).toEqual('ID-S2-1')
+      expect(party.links.SINGLE2[0].linkedPartyId).toEqual('ID-S2-1')
       expect(party.links.MULTI1[0]).toBeDefined()
-      expect(party.links.MULTI1[0].linkedId).toEqual('ID-M1-1')
+      expect(party.links.MULTI1[0].linkedPartyId).toEqual('ID-M1-1')
       expect(party.links.MULTI1[1]).toBeDefined()
-      expect(party.links.MULTI1[1].linkedId).toEqual('ID-M1-2')
+      expect(party.links.MULTI1[1].linkedPartyId).toEqual('ID-M1-2')
     })
 
   })