+++ src/utils/assetManagers/assetManagers.test.js
@@ -7,7 +7,7 @@ api.config({
   stage: 'staging',
   token: process.env.API_TOKEN
 })
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 60000
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 20000
 
 describe('utils/assetManagers', () => {
   describe('_parseAM function', () => {
@@ -70,21 +70,21 @@ describe('utils/assetManagers', () => {
         .catch(error => {})
       expect(promise).toBeInstanceOf(Promise)
     })
-    test('amends', done => {
+    test('amends', async done => {
       let dboi
-      retrieve({ AMId: 4 })
-        .then(res => {
-          dboi = res.defaultBookOwnerId
-          res.defaultBookOwnerId++
-          return amend({ assetManager: res, AMId: res.assetManagerId })
-        })
-        .then(res => {
-          expect(res.defaultBookOwnerId).toEqual(dboi+1)
-          done()
-        })
-        .catch(err => {
-          console.error(err)
-        })
+      let res = await retrieve({ AMId: 4 })
+      if (res.length === 0) {
+        console.error('amend: Results is empty, force fail after timeout.')
+        return
+      }
+      if (res.assetManagerStatus === 'Inactive') {
+        res = await reactivate({ AMId: res.assetManagerId })
+      }
+      dboi = res.defaultBookOwnerId
+      res.defaultBookOwnerId++
+      res = await amend({ assetManager: res, AMId: res.assetManagerId })
+      expect(res.defaultBookOwnerId).toEqual(dboi+1)
+      done()
     })
   })
 
@@ -97,24 +97,21 @@ describe('utils/assetManagers', () => {
 
   describe('reactivate and deactivate', () => {
     // Get an Asset Manager. If it is inactive, reactivate, check then deactivate and check
-    it('reactivates an inactive AM and deactivates an active AM', () => {
-      retrieve({ AMId: 2 })
-        .then(res => {
-          if (res.assetManagerStatus === 'Inactive') {
-            return reactivate({ AMId: 2 })
-          }
-          return res
-        })
-        .then(res => {
-          expect(res.assetManagerStatus).toEqual('Active')
-          return deactivate({ AMId: 2 })
-        })
-        .then(res => {
-          expect(res.assetManagerStatus).toEqual('Inactive')
-        })
-        .catch(err => {
-          expect(err).toBeUndefined()
-        })
+    it('reactivates an inactive AM and deactivates an active AM', async done => {
+      let res = await retrieve({ AMId: 1 })
+      if (res.length === 0) {
+        console.error('reactivate/deactivate: Results is empty, force fail after timeout.')
+        return
+      }
+      if (res.assetManagerStatus === 'Inactive') {
+        res = await reactivate({ AMId: res.assetManagerId })
+      }
+      expect(res.assetManagerStatus).toEqual('Active')
+      res = await deactivate({ AMId: res.assetManagerId })
+      expect(res.assetManagerStatus).toEqual('Inactive')
+      res = await reactivate({ AMId: res.assetManagerId })
+      expect(res.assetManagerStatus).toEqual('Active')
+      done()
     })
   })
 })